refactor : 폼 제출 시 사용되는 함수들을 페이지로 이동
#71
