+++ src/schema/useLoginForm.ts
@@ -16,9 +16,5 @@ export const useLoginForm = () => {
     resolver: zodResolver(schema),
   })
 
-  const handleSubmitLoginForm = (formData: LoginFormType) => {
-    alert(JSON.stringify(formData, null, 2))
-  }
-
-  return { formMethod, handleSubmitLoginForm }
+  return formMethod
 }
